# aplicativos/backend/API/swagger.json
{
    "info": {
        "description": "This is a sample server",
        "title": "Swagger",
        "version": "1.0.0"
    },
    "host": "localhost:4000",
    "basePath": "/v1",
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "Bearer <token>"
        }
    },
    "swagger": "2.0",
    "paths": {
        "/posts/{postId}/comments": {
            "get": {
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "user's id.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "This function get comments of a post",
                "tags": [
                    "Comment"
                ],
                "responses": {
                    "200": {
                        "description": "An array of posts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Comment"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "post": {
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "user's id.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "comment",
                        "in": "body",
                        "description": "the new point",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "description": "This function post a Comment into a Post",
                "tags": [
                    "Comment"
                ],
                "responses": {},
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
        "/feed": {
            "get": {
                "parameters": [],
                "description": "This function get posts",
                "tags": [
                    "Feed"
                ],
                "responses": {
                    "200": {
                        "description": "An array of posts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Post"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/posts": {
            "get": {
                "parameters": [],
                "description": "This function get posts",
                "tags": [
                    "Post"
                ],
                "responses": {
                    "200": {
                        "description": "An array of posts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Post"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "post": {
                "parameters": [
                    {
                        "name": "post",
                        "in": "body",
                        "description": "the new point",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    }
                ],
                "description": "This function Post a Post",
                "tags": [
                    "Post"
                ],
                "responses": {},
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/posts/{id}": {
            "get": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user's id.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "This function to get a post by id",
                "tags": [
                    "Post"
                ],
                "responses": {
                    "200": {
                        "description": "post"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "put": {
                "parameters": [
                    {
                        "name": "post",
                        "in": "body",
                        "description": "the new point",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user's id.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "This function to get a post by id",
                "tags": [
                    "Post"
                ],
                "responses": {},
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user's id.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "This function to get a post by id",
                "tags": [
                    "Post"
                ],
                "responses": {},
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/posts/{id}/like": {
            "post": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user's id.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "\nThis function to get a post by id",
                "tags": [
                    "Post"
                ],
                "responses": {},
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/profiles": {
            "get": {
                "parameters": [],
                "description": "This function list the profiles",
                "tags": [
                    "Profile"
                ],
                "responses": {
                    "200": {
                        "description": "An array of user info"
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/profiles/search?q={q}": {
            "get": {
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "profile id.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "\nThis function to get a post by id",
                "tags": [
                    "Profile"
                ],
                "responses": {
                    "200": {
                        "description": "profile"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/profiles/{id}": {
            "get": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "profile id.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "This function to get a post by id",
                "tags": [
                    "Profile"
                ],
                "responses": {
                    "200": {
                        "description": "profile"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/profiles/{id}/follow": {
            "post": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user's id.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "\nThis function to get a post by id",
                "tags": [
                    "Profile"
                ],
                "responses": {},
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "Comment": {
            "required": [
                "description",
                "profile",
                "post"
            ],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": ""
                },
                "description": {
                    "type": "string",
                    "description": "the comment"
                },
                "profile": {
                    "type": "Profile",
                    "description": ""
                },
                "post": {
                    "type": "Post",
                    "description": ""
                }
            }
        },
...
        "Post": {
            "required": [
                "title",
                "description",
                "profile"
            ],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": ""
                },
                "title": {
                    "type": "string",
                    "description": "title"
                },
                "description": {
                    "type": "string",
                    "description": "description"
                },
                "profile": {
                    "type": "Profile",
                    "description": "profile"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Comment"
                    }
                }
            }
        },
...
    },
...
}